requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
aiohttp-socks==0.6.0
async-timeout==3.0.1
attrs==21.2.0
black==21.9b0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.1
future==0.18.2
h11==0.12.0
h2==4.0.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.2
jsonschema==3.2.0
Logbook==1.5.3
matrix-nio==0.18.6
multidict==5.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.3.0
psycopg2-binary==2.9.1
pycryptodome==3.10.1
pyrsistent==0.18.0
python-socks==1.2.4
regex==2021.8.28
requests==2.26.0
six==1.16.0
tomli==1.2.1
typing-extensions==3.10.0.2
unpaddedbase64==2.1.0
urllib3==1.26.7
yarl==1.6.3
